requirements.txt
boto
PyYAML